refactor: moved components into /projects and /posts for better organization
