+++ frontend/src/App.js
@@ -26,7 +26,7 @@ import TitleButton from './components/navigation/TitleButton.js'
 import RequireFacultyProfile from './components/authentication/RequireFacultyProfile.js'
 import FacultyProfileView from './components/profiles/FacultyProfileView.js'
 import FacultyProfileEdit from './components/profiles/FacultyProfileEdit.js'
-import ResearchOpportunityForm from './components/ResearchOpportunityForm.js'
+import ResearchOpportunityForm from './components/projects/ResearchOpportunityForm.js'
 import AdminFacultyEdit from './components/admin/AdminFacultyEdit.js'
 import ManageVariables from './components/admin/ManageVariables.js'
 import AdminStudentEdit from './components/admin/AdminStudentEdit.js'
