+++ frontend/src/__tests__/posts/MainAccordion.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import MainAccordion from '../components/MainAccordion'
-import { errorLoadingPostingsMessage, NO_MAJORS_MSG, viewFacultyFlag, viewStudentsFlag, viewProjectsFlag } from '../resources/constants'
-import { mockResearchOps, getAllFacultyExpectedResponse, mockStudents } from '../resources/mockData'
+import MainAccordion from '../../components/posts/MainAccordion'
+import { errorLoadingPostingsMessage, NO_MAJORS_MSG, viewFacultyFlag, viewStudentsFlag, viewProjectsFlag } from '../../resources/constants'
+import { mockResearchOps, getAllFacultyExpectedResponse, mockStudents } from '../../resources/mockData'
 
 describe('MainAccordion', () => {
   test('renders error message when no postings are provided', () => {
