+++ frontend/src/components/navigation/SharedLayout.js
@@ -12,7 +12,7 @@ import { useNavigate, Link } from 'react-router-dom'
 import TitleButton from './TitleButton'
 import SearchBar from '../filtering/SearchBar'
 import ViewProfile from '../profiles/ViewProfile'
-import PostsLayout from '../PostsLayout'
+import PostsLayout from '../posts/PostsLayout'
 import { bgColor } from '../../resources/constants'
 import Sidebar from '../filtering/Sidebar'
 import PropTypes from 'prop-types'
