+++ frontend/src/components/posts/MajorAccordion.js
@@ -7,9 +7,9 @@
 import React from 'react'
 import { Accordion, AccordionSummary, AccordionDetails, Typography, Box } from '@mui/material'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import PostList from './posts/PostList'
+import PostList from './PostList'
 import PropTypes from 'prop-types'
-import { viewProjectsFlag } from '../resources/constants'
+import { viewProjectsFlag } from '../../resources/constants'
 
 function MajorAccordion ({ major, numPosts, setSelectedPost, isStudent, isFaculty, isAdmin, postsView }) {
   return (
